+++ src/nestjs/healthcheck/index.ts
@@ -5,5 +5,6 @@
 export const HEALTHCHECK_ENDPOINT = 'health';
 
 export { BaseHealthIndicatorService } from './base-health-indicator.service.js';
+export type { HealthChecker } from './checker.js';
 export { HealthCheckModule } from './module.js';
 export * from './terminus.module.js';